Fixed bug in oldValueIteration

+++ AA/src/autonomousagents/actions/Action.java
@@ -5,9 +5,9 @@ import autonomousagents.Point;
 
 public abstract class Action
 {
-	private float probability;
+	private double probability;
 
-	public Action(final float probability)
+	public Action(final double probability)
 	{
 		this.probability = probability;
 	}
@@ -28,12 +28,12 @@ public abstract class Action
 	 */
 	public abstract Point apply(final Point p);
 
-	public float getProbability()
+	public double getProbability()
 	{
 		return this.probability;
 	}
 
-	public void setProbability(final float probability)
+	public void setProbability(final double probability)
 	{
 		this.probability = probability;
 	}
